+++ code/http/HttpRequest.cpp
@@ -87,7 +87,7 @@ HttpRequest::HTTP_CODE HttpRequest::parse(Buffer &buffer)
 
 HttpRequest::HTTP_CODE HttpRequest::parse_request_line(const std::string &line)
 {
-    std::regex patten("^([^ ]*) https?:\/\/([^ ]*) HTTP\/1.1$");
+    std::regex patten("^([^ ]*) ([^ ]*) HTTP/1.1$");
     std::smatch subMatch;
     if(std::regex_match(line, subMatch, patten)) 
     {   
@@ -105,19 +105,19 @@ HttpRequest::HTTP_CODE HttpRequest::parse_request_line(const std::string &line)
         }
 
 
-        std::string path = subMatch[2];
-        /* 去处url中ip:port部分 */
-        size_t rootPos = path.find("/");
-        if (rootPos == std::string::npos)
-            return BAD_REQUEST;
-        m_url = m_url.substr(rootPos);
+        // std::string path = subMatch[2];
+        // /* 去处url中ip:port部分 @FIXME 去除https*/
+        // size_t rootPos = path.find("/");
+        // if (rootPos == std::string::npos)
+        //     return BAD_REQUEST;
+        m_url = subMatch[2];
 
         /* 当url为/时，显示判断界面 */
         if (m_url.compare("/") == 0)
         {
             m_url = "/judge.html";
         }
-
+        
         m_version = "HTTP/1.1";
         m_check_state = CHECK_STATE_HEADER;
         return NO_REQUEST;
