+++ root/judge.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <title>WebServer</title>
+    </head>
+    <body>
+    <br/>
+    <br/>
+    <div align="center"><font size="5"> <strong>welcome</strong></font></div>
+	<br/>
+		<br/>
+		<form action="0" method="post">
+ 			<div align="center"><button type="submit">new user</button></div>
+                </form>
+		<br/>
+                <form action="1" method="post">
+                        <div align="center"><button type="submit" >login</button></div>
+                </form>
+		
+		
+        </div>
+    </body>
+</html>
+
+
